Major changes - PG and PF scenarios; BID global
P_Hydro and P_Grid are now scenario-dependent variables and the sum this
variables, in each scenario, must be equal to a new global variables:
BID.

Removed the constraint defining the storage final value in the day-ahead
and operational strategy.

Expeted Value and Point Forecast functions are operational, but the
others are not.